+++ src/content/index.ts
@@ -20,7 +20,7 @@ import { refreshMainTitle } from './titles/mainTitle';
 import { refreshDescription } from './description/MainDescription';
 import { handleAudioTranslation } from './audio/audioIndex';
 import { handleSubtitlesTranslation } from './subtitles/subtitlesIndex';
-import { maybeShowSupportToast } from './support/AskForSupport';
+import { maybeShowSupportToast } from './SupportToast/toast';
 
 
 coreLog('Content script starting to load...');